fix: cast

+++ test/external/external_benchmark_openpilot.py
@@ -30,7 +30,7 @@ if __name__ == "__main__":
   ort_out = onnx_output[0]
 
   run_onnx = get_run_onnx(onnx_model)
-  tinygrad_out = next(iter(run_onnx(new_inputs).values())).numpy()
+  tinygrad_out = next(iter(run_onnx(new_inputs).values())).cast(dtypes.float32).numpy()
 
   # validate
   np.testing.assert_allclose(ort_out, tinygrad_out, atol=2e-3, rtol=1e-2)
@@ -40,7 +40,7 @@ if __name__ == "__main__":
   tms = []
   for _ in range(10):
     st = time.perf_counter_ns()
-    ret = next(iter(run_onnx(new_inputs).values())).numpy()
+    ret = next(iter(run_onnx(new_inputs).values())).cast(dtypes.float32).numpy()
     tms.append(time.perf_counter_ns() - st)
   print(f"unjitted: {min(tms)*1e-6:7.2f} ms")
 
@@ -48,6 +48,6 @@ if __name__ == "__main__":
   run_onnx_jit = TinyJit(run_onnx)
   for _ in range(10):
     st = time.perf_counter_ns()
-    ret = next(iter(run_onnx_jit(new_inputs).values())).numpy()
+    ret = next(iter(run_onnx_jit(new_inputs).values())).cast(dtypes.float32).numpy()
     tms.append(time.perf_counter_ns() - st)
   print(f"jitted: {min(tms)*1e-6:7.2f} ms")
